Update test to ensure max size uint literals can be assigned
[Reviewed by vass]

This patch changes the existing test uint8-64.chpl to better check the
bounds of uints. It replaces some of the assignment values with ones that
are the maximum integer for each size. The older checks were not ensuring
that you could actually set the maximum size with a literal. For example
the old check of 'ui(32)=2000000000 is 2000000000' did not detect that you
could not assign a literal value of max(int(32))+1 or greater to a
uint(32). The new form should ensure that this doesn't happen in the
future.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23295 3a8e244f-b0f2-452b-bcba-4c88e055c3ca