Fix a bug that prevented different reduction strategies from working in genGraphs
[Reviewed by vass]

When I initially committed the means to use different reduction strategies from
the .graph files, I left off a split. Causing the generate option list to end
up looking like ['m','i','n','m','a','x', ...]. This commit fixes that issue
and adds a check that will throw a warning if the parsed generate option is not
recognized.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23299 3a8e244f-b0f2-452b-bcba-4c88e055c3ca