Disable sync type compiler-inserted reads when using minimal modules
[Reviewed by David]

This patch disables insertion of the _statementLevelSymbol() function
during normalize and also in insertReturnTemps() during resolution
when compiling with --minimal-modules.  This function is used to
implement the semantics of sync types in the case of variables whose
values are not captured, but it is not needed for --minimal-modules as
sync types are not supported for minimal modules.

While there clarified comments regarding why we do this (including in
ChapelBase.chpl where _statmentLevelSymbol() is defined) and fixed a
type-o in buildDefaultRunctions.cpp regarding --minimal-modules
compilation.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23301 3a8e244f-b0f2-452b-bcba-4c88e055c3ca