Change default stack size test
[ok'd by greg]

test/exercises/RandomNumber6.chpl was recently failing for qthreads and numa,
so guard pages were turned off thinking that fixed the problem.  However, the
problem was still intermittent and the reason ended up being that this test
creates 10,000 tasks and with the default stack size of 8MB that means qthreads
could have ~78GB of memory allocated just for the tasks' stacks.  The reason it
was intermittent was because there is a sleep right after the tasks are
created, and depending on how quickly the tasks are created vs how quickly the
sleep is over and the tasks do there work and are destroyed determines the max
amount of memory used at any point. 

This patch lowers the default stack size for this test and removes the
explicitly setting of guard pages. 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23307 3a8e244f-b0f2-452b-bcba-4c88e055c3ca