Enables testing of C code via the normal start_test/sub_test system
[Reviewed by hilde]

This commit makes it so that files ending with .test.c are counted as
tests alongside .chpl files. This extension was chosen over plain .c since
we already use .c files in various places in the testing system. The C
compiler, along with default arguments, is found via `compileline
--compile`. By doing this, it is easy to use headers in the runtime.

To help with getting .c files from the runtime, the ability to place
environment variables in .compopts (and others) was added via
os.path.expandvars.

An example compopts file to use the runtime timers is:
     $CHPL_HOME/runtime/src/chpl-timers.c

This commit does not contain any tests using this functionality, but a
future commit I am preparing will.

A check around the file I/O inside of cleanup was added as well.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23308 3a8e244f-b0f2-452b-bcba-4c88e055c3ca