Change the default stack size for mallocOutOfMemory test
[ok'd by greg]

test/memory/shannon/outofmemory/mallocOutOfMemory artificially sets the virtual
memory limit before calling sub_test so that we can see if we get reasonable
error messages when we run out of memory. An existing problem is that this
limit gets set for compilation as well. It has been noisy in other
configurations because of this reason. 

For qthreads however, this test was failing at runtime startup, not compile
time. The reason for this is because currently qthreads preallocates stack
space for 128 threads. This amount of memory is small for qthread's default
stack size, but since we set the default stack size to 8MB we allocate ~1.1GB
which exceeds the current ~400MB ulimit set. 

One of the options was to bump up the ulimit again, but the 128 tasks might be
per sublocale so this test could introduce more noise in the future for
qthreads if that was changed. Instead this patch just lowers the stack size for
the test, bringing the startup memory to well below the ulimit set, and
allowing the test to run properly. 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23309 3a8e244f-b0f2-452b-bcba-4c88e055c3ca