Add configuration searching to explicitly named .good files
[reviewed by sung]

Previously if a .good file was explicitly named either in a .compopts or a
.execopts that file, and only that file were considered when searching for a
.good file to compare to. 

This patch changes it so that if you had an explicit .good file like
test.option=1.good, it will also search for test.option=1.lm-numa.good, or
test.option=1.no-local.good and so on for the different configurations that are
checked. 

The motivation behind this was test/distributions/sungeun/dataPar2.chpl where
there are multiple compopts and multiple execopts and each execopts has an
explicit .good file specified. These .good files differ for numa, but there
wasn't an existing way to find different files based on locale model for
explicitly specified .good files. Even if you got rid of the explicit names,
and named the .good files after the compopt and execopt num you still wouldn't
have a way to specify different.good files for different configurations such as
numa. 

To do this rather than add this ability to both the compiler and execution
.good search logic, I combined that into a common function. This should make it
easier make changes to what types of .good files we look for in the future. 

This should preserve all existing .good searching behavior. 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23310 3a8e244f-b0f2-452b-bcba-4c88e055c3ca