Look for --no-local in globalCompopts and envCompopts
[not reviewed, trivial]

r22338 changed the contents of globalCompopts to not include those coming from
the COMPOPTS environment variable. That was where --no-local was coming from so
with that change, sub_test was unable to find no-local .good files. This patch
just fixes the search for --no-local, I'm not sure if theres other things
that were affected. 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23311 3a8e244f-b0f2-452b-bcba-4c88e055c3ca