Introduced initial support for finding the subdomain of an array for the current locale.
The current proposal is to eventually provide the following methods,
grouped by domain map/domain/array...

    dm.targetLocales()                  // the locale grid
    dm.targetLocDom()                   // the domain of the locale grid
    dm.hasSingleLocalSubdomain() param  // can the local subdomain be
                                        // represented as a single domain?
                                                            -----------

    D.targetLocales()            // same as above
    D.targetLocDom()             // same as above
    D.hasSingleLocalSubdomain()  // same as above

    D.localSubdomain()   // the subdomain for the current locale,
                         // defined only if D.hasSingleLocalSubdomain()
                         // note: non-trivial for Block

    D.localSubdomains()  // iterator - yields the subdomain(s)
                         // for the current locale
                                                                                                                                                               -----------

    A.targetLocDom()              // same as above
    A.targetLocales()             // same as above
    A.hasSingleLocalSubdomain()   // same as above
    A.localSubdomain()            // same as above; trivial for Block
    A.localSubdomains()           // same as above

    // continuing with the theme
    A.localSubarray()    // alias to sub-array for the current locale
    A.localSubarrays()   // iterator, yields aliases(??)
                         // to sub-array(s) for the current locale

Comments from Brad:

One open question is whether to extend queries that are logically
related to the distribution or domain to the domain and/or array as a
convenience vs. requiring the user to say something like A.dom.dist...
From a user perspective it seems nice; from a developer/minimalist
perspective, it seems a little annoyingly redundant.

It's worth noting that while many of the queries can forward (like
targetLocales()), others may not be able to, or else will need to be
conservative.  E.g., in block-cyclic, no locale owns a single subdomain
of the distribution as a whole, but for a given domain, it might (if the
domain is small enough).  But we can't know this at compile-time, which
might suggest we should be conservative. [This depends much on whether
D.hasSingleLocalSubdomain() is a param function.]

Along those same lines, I don't think it makes sense to support queries
relating to sublocales on a distribution since distributions don't have
sublocales as part of their user-level definition, at least in my
opinion (even though they may happen to use them as part of their implementation).

----------------------------------------

Open Question: 
    should there be the param method dsiHasSingleLocalSubdomain()
    and/or non-param method dsiLocalSubdomain() (generating
    a compiler error "should not call this") on BaseArr and/or
    BaseDom/BaseDist ?


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23312 3a8e244f-b0f2-452b-bcba-4c88e055c3ca