Switch order of module removal from global lists
[trivial not reviewed]

I was removing modules from userModules/allModules/mainModules after the call
to clean_gvec for gModuleSymbols occured, meaning that I was checking isAlive
on modules that were already deleted. This just switches the order so that we
remove modules from the user/all/main lists before deleting the modules AST and
removing it form gModuleSymbols.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23319 3a8e244f-b0f2-452b-bcba-4c88e055c3ca