Re-search libchpl.a after libqthreads_chpl.a.
[Reviewed verbally by Sung.]

With --minimal-modules, at the time we link libchpl.a we have no
outstanding references to any symbols defined in libchpl.a:chpl-tasks.o,
so we don't include that in the program.  Later we link in the qthreads
tasking code from libqthreads_chpl.a.  That references symbols in
libchpl.a:chpl-tasks.o, but by then it's too late to bring it in because
(omitting some details) the linker only does one left-to-right pass of
the files on its command line.  To fix this, search libchpl.a again
after linking in libqthreads_chpl.a.  (This problem is peculiar to the
qthreads tasking layer, for which the tasking layer code that would
otherwise be in the Chapel runtime library is in this third-party
library instead.)


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23320 3a8e244f-b0f2-452b-bcba-4c88e055c3ca