Added date-based sorting to genGraphs.
[Contributed by Kyle; reviewed by Vass]

Sorting allows genGraphs to handle .dat files where lines for a given date
are not necessarily consecutive and/or not in the order of date increase.
Prior to that change the resulting graph would still contain all the data
points, but they would not be displayed and/or aggregated nicely.

Sorting is done twice: first, the lines are textually sorted. This ensures
the "consecutive" property. However, the dates can still be out of order,
e.g. 05/10/13 will go after 02/10/14.

The second sorting is on the list 'df.lines'. That orders the list entries
by date. NB a list entry looks like this, i.e. it is itself a list of lists:

  [[date-for-this-entry, first-line-data1, first-line-data2, ...],
   [date-for-this-entry, second-line-data1, second-line-data2, ...],
   ...
  ]

So sorting is done with the first key being the entry's first element
of the first element.

While there, I added minor comments.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23322 3a8e244f-b0f2-452b-bcba-4c88e055c3ca