Add support to pass array arguments to extern'ed functions
This commit adds support for array arguments to extern'ed functions by
adding a new pass to the compiler, expandExternArrayCalls.

The goal of this expandExternArrayCalls is to convert any extern proc
declarations that contain an array argument to take a c_ptr instead and
provide a wrapping function that can handle being passed a Chapel array.

So we take a declaration like:

   extern proc print_array(x: [] int, n: int);

And create:

    // This will have the name print_array in the generated C code
    extern proc chpl__extern_array_print_array(x: c_ptr(int), n: int);

    // This will have the name chpl__extern_array_wrapper_print_array in
    // the generated C code
    inline proc print_array(x: [] int, n: int) {
      chpl__extern_array_print_array(c_ptrTo(x), n);
    }

If no element type is specified for the array a c_void_ptr will be used.
If we leave out the element type in the example above the result would be:

    extern proc chpl__extern_array_print_array(x: c_void_ptr, n: int);

    inline proc print_array(x: [], n: int) {
      chpl__extern_array_print_array((c_ptrTo(x)):c_void_ptr, n);
    }


Two tests will fail after this commit:

test/performance/compiler/elliot/passCheck
test/users/bugzilla/bug794131/bug794131

The passCheck failure will be resolved after committing.

The bugzilla test seems like it should be removed. It can't actually
compile and was depending on failing due to having an array argument to
an extern'ed function, but it is not a future.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23324 3a8e244f-b0f2-452b-bcba-4c88e055c3ca