Rework computing number of threads in tasks=qthreads.
[Reviewed by Elliot.]

Redo how we compute the number of worker threads in the Qthreads tasking
layer.  In order of precedence:
- Use the CHPL_RT_NUM_THREADS_PER_LOCALE value, but don't let that
  oversubscribe the hardware.
- Use QTHREAD_HWPAR, and do allow oversubscribing.
- Use the Qthreads default (which won't oversubscribe).

In any case, silently limit the number of worker threads to the comm
layer's maximum thread count, if there is one.

The need for this was demonstrated when we discovered that even when we
set QT_WORKER_UNIT=core, we still got a worker per PU.  We eventually
found that this was due to our expressly setting the number of threads
to the value returned by chpl_numCoresOnThisLocale() (which actually
returns the number of PUs, not cores) even in cases where we didn't need
to do anything.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23328 3a8e244f-b0f2-452b-bcba-4c88e055c3ca