Updates debugDataPar output and testing
[reviewed by sung]

The debug data par output for numa was never running correctly. There were a
couple of reasons for this: first because we never threw -sdebugDataParNuma,
second because we didn't have a way of creating locale model specific .good
files and lastly because the debug output is non-deterministic and we didn't
output it in a very sortable manner. 

This changes the debugDataParNuma output to be sortable, and more consistent
with debugDataPar output and updates debugDataPar to print out some additional
information that debugDataParNuma was. 

This patch also updates testing for debugDataPar to expect the additional debug
output and adds .good files for numa. For numa QT_NUM_SHEPHERDS is set to 2 to
force the test to run with 2 sublocales so that the execution path for numa is
consistent regardless of the machine its running on. Without this the output
would be drastically different on machines that have various numbers of
sockets/numa domains. 

test/functions/iterators/sungeun/forwarding/forwardDefault* have skipifs for
numa. The reason for the skipifs despite the fact that this also adds .good
files for those tests under numa is that the tests change ignoreRunning and
right now numa is not task count sensitive for sublocales so the output is
still nodeterministic. The .good files are added anyways, becuase they will the
correct .good files when the skipif can be removed (and if not they will be
really close)


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23333 3a8e244f-b0f2-452b-bcba-4c88e055c3ca