Added a prediff for the future users/ferguson/extern_string_test.chpl
While second line in the output of the test is non-deterministic,
most of the time it is "returned string arg r". This has been in .bad.

Occasionally, however, there is some stuff at the end of that string,
most recently five \0-characters. I am adding a prediff and modifying
the .bad so that that case does not trigger the ".bad mismatch" error.

It may be that the test will still produce something nondeterministic
that my prediff will not handle. We can deal with that when it happens.

My intention for the prediff is that if the output of the test is correct
(i.e. what's in .good), the prediff does not change that.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23337 3a8e244f-b0f2-452b-bcba-4c88e055c3ca