Clean up testVFlag and testVerboseFlag.
[Reviewed by Tom.]

Over time multilocale/numLocales/bradc/testVFlag and testVerboseFlag
have gotten messy.  Each has a .prediff to build a .good file to compare
against and adjust the output from the test before doing that compare,
but in some ways the .prediffs were making choices based on the wrong
configuration variable and in one case (comm=none) the .good file we'd
gone to the trouble of building was then ignored (see below).

This change cleans all this up a little.  There are still .prediff files
(which are now identical for the two tests, although nothing depends on
that).  These still build a .good file containing an optional launcher
command, the "executing on" output from the comm layer, and the expected
output from the program.  But, how this is done is now quite different:

We get the launcher command from .launcher-$CHPL_LAUNCHER.goodstart,
with the string 'EXECOPTS' replaced with the execution options stored in
$EXECOPTS by start_test.  For CHPL_LAUNCHER=none, this file is empty,
because we don't have a launcher command.  (In the past we used to get
the launcher binary name from a .comm-$CHPL_COMM.goodstart file, and
then the .prediff built the launcher command itself.  This worked as
long as there was a tight binding between comm layers and launchers, but
that hasn't been true for quite a while, at least on the Cray systems.)

We get the "executing on" line (which is printed by the comm layer) from
.comm-$CHPL_COMM.goodcont, with the string 'UNAME' replaced with the
output from a 'uname -n' command.  These .goodcont files are all new.
The "executing on" line used to be built by the .prediff script itself,
except in the case of comm-none, where it came from the .goodstart file.
But for comm=none, we ended up ignoring the built .good anyway, because
there was an overriding .comm-none.good already in the repository.  This
overriding .comm-none.good file is deleted by this commit, since the
.prediff can now do the right thing for comm-none.

We get the expected output from the .goodstop file.  This didn't change.

Finally, I also rearranged the logic used to adjust the .exec.out.tmp
contents prior to comparing with the .good, to clarify the relationship
between the adjustments we perform and the configuration settings that
result in our needing to do them.  This happens in a series of case
statements based on CHPL_TARGET_PLATFORM, CHPL_TASKS, and CHPL_LAUNCHER.

With these changes, testVFlag and testVerboseFlag work with all of the
following configuration settings:

  TARGET    LAUNCHER    COMM      TASKS
  ------    --------    ----      -----
  linux64   none        none      fifo, massivethreads, muxed, qthreads
            amudprun    gasnet    fifo, massivethreads, muxed, qthreads
  cray-xe   none        none      fifo, muxed, qthreads
            aprun       gasnet    fifo, muxed, qthreads
                        ugni      fifo, muxed, qthreads
  
(I didn't test massivethreads on the Cray system because I had trouble
building it due to an unrelated static/dynamic linking mismatch.)


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23341 3a8e244f-b0f2-452b-bcba-4c88e055c3ca