Added -genGraphOpts to start_test.
[reviewed by Elliot]

Allow the user to pass options to genGraphs when running start_test.
This would be useful, for example, for genGraphs --numericX (r23339).

Note: only a single 's' in -genGraphOpts.

At first I wanted to add -numericX to start_tests, but then
I chose -genGraphOpts as a more general solution,
for example the user may want to pass the end date (-e).

An addendum to r23339/r23340, thanks to Elliot:
If the user uses genGraphs --numericX for numbers other than
revision numbers, they will still get release markers on their graphs.
We decided to leave it alone for now and reconsider if+when
it becomes an issue.

While there - added quotes in ''switch ( "$argv[1]" )''
- to allow space-containing options.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23342 3a8e244f-b0f2-452b-bcba-4c88e055c3ca