Add feature request future: A more robust module system
from use_shadows_formal.future:

Currently the only way we have to reference another module outside of the
current file is a 'use' statement. There are a few problems with this. One of
the primary issues is shown in this test. We 'use' a module that defines one
symbol (bar) inside of the procedure test. Now because we 'use' all of the
standard modules inside of every other module, when we go to compile we will
get an error since there is a function inside of Sys named addr that will
shadow our formal.

A method of allowing other modules to be used with out dumping a very large
number of symbols into the local context should be provided. In addition a
method to control which symbols are imported and exported is also needed.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23349 3a8e244f-b0f2-452b-bcba-4c88e055c3ca