Add new bug future for functions that are converted to methods not being scoped properly
This commit separates this bug out from the 'use_shadows_formal' issue.

from proc_scoping.future:

In this test, the function int.print_me is implicitly converted to a
method of the form `print_me(MethodToken, this)`. In doing so it loses
it's original scope of being a function on the type 'int'. This causes it
to shadow the formal print_me when it should not.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23351 3a8e244f-b0f2-452b-bcba-4c88e055c3ca