Clean up paratest.server and re-enable timeouts.
There was a problem with the previous implementation.  Each time free_workers()
is called, we issue a "wait;" for each "Feed me" file found in the .synch
directory.  The code to feed workers (while untested directories remain)
unlinks the "Feed me" file when the worker is fed [l.231].  

In contrast, the code that waits for all workers to finish does not remove
"Feed me" files.  This results in "wait;" being issued multiple times for some
workers, and that causes the "wait;" to block.  (The "wait;" statment waits for
an unspecified child process to finish; it cannot know that the one we meant
has already finished.  Using waidpid() would be one way to solve this
problem.)

This patch resolves the problem


DETAILS

I adjusted whitespace on lines 172..184 and 260..265.

I added code to remove .synch directory entries for workers after there is no
more work, so we don't end up waiting multiple times for a child process to
terminate.  In theory, this calls "wait" only once per spawned process and thus
allows free_workers() to be non-blocking.

That in turn allows final worker timeouts to be checked once per second, thus
restoring the worker timeout feature.  Using that feature means that
paratest.server *will terminate* and print the usual test summary.  In the case
that one or more workers hung, their IDs are also printed.

[tested on examples (-nodepara 1)]
[tested by Chris Wailes]
[reviewed by vass]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23358 3a8e244f-b0f2-452b-bcba-4c88e055c3ca