Remove externFloat4calls.future and externFloat4calls.someFields.future.
These feature requested in these futures was provided long ago, but the tests
continued to fail due to a formatting error in the .good file.  (Here is a place
where having .bad files might have helped.)

float4.c was updated so it prints a newline at the end of the output.
The files externFloat4calls.good and externFloat4calls.someFields.good were
updated to reflect the actual (expected and correct) output.
The corresponding .future files were removed.

[tested (std)]
[tested (no-local)]
[reviewed by vass]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23360 3a8e244f-b0f2-452b-bcba-4c88e055c3ca