Fix for calling constructors through typedefs.
This patch resolves the feature request
  test/types/typedefs/lydia/def-new-on-type-alias.future
.  

Though this feature has merit by itself, it also supports calling
the default constructor for record variables (or fields) created without an
explicit initializer.  In many test cases, the var (or field) is declared using
a typedef as the type expression.  Creating a call to the default constructor
using text substitution (as is currently done) does not work, because the
resolved type of the typedef has a different name than the typedef itself.

At present, the normalize pass converts a type alias into a
chpl__typeAliasInit() expression, which is resolved and then thrown away during
resolution.  The main reason for creating and resolving this expression is to
issue an error if an attempt is made to initialize a type variable using a
value that known only at run-time.

The main change involves also leaving the definition of a typedef in the "init"
field of the containing DefExpr.  This definition may be used during resolution
to resolve the base type of expressions that look like constructor calls and
were not resolved in the normalized pass.  (That code path is not currently
used, but is provided to support future work, as described above.)  The "init"
fields are removed after resolution (since all types should be resolved by the
time resolution is done), rather than during normalization.  

Note that the DefExprs themselves now persist to the end of compilation because
they are no longer removed in normalization.  This is required, since a typedef
may be used without being resolved (e.g. in calling an extern function).  Since
typedefs can be scoped inside functions, we need the new tests in
copyPropagation to ignore them.

In call_constructor_for_class, I use the stashed definition of a type alias to
resolve the type that is being used for the constructor call.  This corrects
the long-standing bug that prevented a user from calling a constructor through
a typedef.


TODO: Add a verification check for why unresolved symbols remain in the tree
after resolution, and get rid of these.

DETAILS

Index: test/types/typedefs/lydia/def-new-on-type-alias.future
Index: test/types/typedefs/lydia/def-new-on-type-alias.good
Resolved this future.  Constructor call syntax with a typedef'ed type name now
resolves to the base type.
The .good file needed an EOL character.

Index: test/types/typedefs/bradc/newTypedef.future
This future is resolved.

Index: compiler/main/checks.cpp
Moved the check that all aggregate types have default initializers and default
constructors here from call_constructor_for_class() in normalize.cpp.

Index: compiler/passes/normalize.cpp
Added a clause to checkUseBeforeDef, so type aliases are treated as definitions
(always defined).  If a type variable is not actually compile-time defined, an
error will be generated during resolution.
In normalize_returns(), indented and reformatted the code for initializing the
return value variable in the case that the function has a declared return type.
In call_constructor_for_class(), resolve type symbols through type aliases
(recursively) by calling the new routine resolveTypeAlias().  Moved the
consistency checks to checks.cpp (see above).
In fix_def_expr(), rather than removing the "init" field of a DefExpr used to
define a type definition, we leave the definition in place by copying it.  This
allows references to the type variable to be resolved through those init fields.

Index: compiler/resolution/functionResolution.cpp
1. Factored out lvalueCheck().
2. In preFold(), added comments.  Also, if the LHS of a move is set previously,
further attempts to set it are ignored.  See paragraph 3 below.
3. Factoring of resolveExpr() out of resolveBlock() causes resolution of param and
non-param functions to proceed using the same control logic.  As a result, in
constructs like:
proc foo() param {
  if (cond) return true;
  return false;
}
both clauses are evaluated.  If cond was true, this results in an attempt to
set the value of the return value variable twice.  I rewrote param
resolution so that the first assignment is the winner.  A better approach might
be to split apart param resolution because different control structures really
are needed for the two kinds of resolution.  But that effort seemed much larger
than the bug at-hand.
4. Factored resolveExpr() out of resolveBlock(), so it could be called
recursively from the new clause added to resolve the definition of typedefs.
5. Added a clause to resolve the init field in a typedef.  This mainly removes
unresolved symbols, but will also generate errors if the type expression cannot
be evaluated.
6. Added a clause to do late resolution of constructor calls.  This is just a
stub, to be expanded when work continues on calling default constructors for
variables and fields.
7. Added a new routine removeInitFields() to remove the definitions of typedefs
after function resolution is complete.  

Index: compiler/optimizations/copyPropagation.cpp
Some typedefs now survive in the tree to copyPropagation, so clauses were added
to handle these properly (they are ignored).  An alternative solution would be
to move the type definitions up to module scope, so they do not appear in the
body of any function.  However, the current way is correct and straightforward.


[tested on / (std)] -- 0 failures
[tested on / (no-local)] -- 0 failures
[tested on / (verify)] -- 0 failures
[reviewed by vass]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23363 3a8e244f-b0f2-452b-bcba-4c88e055c3ca