Minor cleanup for baseAST.h and then fixes for "collatoral damage"
The interpreter will be working on the AST and so I am digging in to
the design of the AST.  Noticed some legacy coding in include/baseAST.h
and decided to do a little house cleaning

a) Removed nested includes that baseAST.h didn't directly rely on

b) Determined that a few fields of BaseAST could be flagged as private or protected

c) The AstTag enum included "sentinel definitions" to support some psuedo-subtype
tests.  These sentinel definitions would never appear in an instance of BaseAST but
caused compiler warnings if one wrote a switch() over the AstTag value.  I've
removed the sentinels.  Tom and I share a view that it might be appropriate to
deprecate the Enum itself in the fullness of time

d) Moved some code from BaseAST.h down in to BaseAST.cpp


The remaining files were touched to address the fallout of these
innocuous changes in baseAST.h.  Primarily changes to add #includes
that had been inherited via baseAST.h or to add statments of the
form "class Expr" in preference to includes.

Ran the test suite and code reviewed by Tom.





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23364 3a8e244f-b0f2-452b-bcba-4c88e055c3ca