Adding tests to make sure remote value forwarding maintains locale of forwarded values (including future for strings)
These tests declare variables of basic types (also in tuples, records,
and classes) on one locale and access them on another including going
back to the original locale.

The string test does not work, though it is not due to remote value
forwarding.  String accessed on remote locales are sometimes copied
even when just their locale ids are accessed.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23366 3a8e244f-b0f2-452b-bcba-4c88e055c3ca