First pass on cleaning up include/{expr.h, symbol.h, type.h}
Additional cleanup for the AST interfaces.

Annotated the classes Expr, Symbol, and Type with "protected" and "private"
to the small extent possible right now.

Added a const qualifier in a few places.

Removed extraneous #include from each header file.

Changes in the implementation files were minimum to match the headers.

Reviewed by Tom.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23370 3a8e244f-b0f2-452b-bcba-4c88e055c3ca