Update class assignment to use PRIM_ASSIGN in place of PRIM_MOVE.
The main change replaces
  { __primitive("move", a, b); }
with
  { __primitive("=", a, b); }
in
   inline proc =(ref a, b: a.type) where isClassType(a.type)
and
   inline proc =(ref a, b:_nilType) where isClassType(a.type)
.  The remaining changes let PRIM_ASSIGN work with combinations of argument
types not previously supported.

The key change is in codegenAssign().  Support was added for the assignment of
'nil' to a wide class pointer (FLAG_WIDE).  To obtain correct code (and keep
the backend C compiler happy), 'nil' must be cast to a wide class object
(FLAG_WIDE_CLASS) of the correct type.  That is it must be the value type of
the wide class reference type of the LHS argument.

The special case clause for PRIM_MOVE in makeHeapAllocations() near line 930
is subsumed by the "else" clause that follows, so it was removed.

[tested on / (std)]
[tested on / (--no-local)]
[reviewed by kbrady]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23375 3a8e244f-b0f2-452b-bcba-4c88e055c3ca