Add --cache-remote compile flag
[Reviewed by Sung and Greg]

This patch adds a --cache-remote compiler flag to enable the cache for remote
data.  It also adds CHPL_CACHE_REMOTE as a global variable available to the
runtime.  In the Chapel modules, whether or not the compile has the
 --cache-remote flag can be tested with param CHPL_CACHE_REMOTE:bool. At
run-time, CHPL_CACHE_REMOTE is a const int containing 1 or 0.

The --cache-remote flag does nothing currently; a future commit will add a
runtime component which the flag will enable.

This patch is part of a series for adding a cache for remote data.
The complete set has been tested extensively. This particular patch
has been double-checked with make check. The new flag will be tested
more once it is connected to the new functionality.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23384 3a8e244f-b0f2-452b-bcba-4c88e055c3ca