Factor out reference counting code, relax
[Reviewed by Sung and Greg]

Before this patch, ChapelDistribution.chpl implements reference counting for
domains, arrays, distributions.  In this patch, we factor out reference
counting methods in general to a new record type, atomic_refcnt, and
implement it in a manner th carefully limits the memory fences (which will be
more important with the cache). In addition, we add proc atomic_fence which
will do barriers on the remote data cache in addition to local processor
barriers.

Sung requested that I make the relaxed consistency only apply when the cache
is enabled; so reference counting behaves more simply when it is not enabled.

Updated two tests that explicitly go in and use reference counts.

Example changes:
 _arrCnt: atomic_int64 -> atomic_refcnt;
 _arrCnt.add(cnt) -> _arrCnt.inc(cnt);
 _arrCnt.fetchSub(1)-1 -> _arrCnt.dec();

This patch is part of a series that has been extensively tested.
This patch individually has been tested with make check (local)
and also that Hello World builds and runs with --no-local.

Future Work:
 Arguably the new module should be called RefCnt.chpl and not
 AtomicsCommon.chpl...



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23385 3a8e244f-b0f2-452b-bcba-4c88e055c3ca